requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
frozenlist==1.4.0
idna==3.4
mando==0.7.1
multidict==6.0.4
numpy==1.25.1
openai==0.27.8
pandas==2.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
radon==6.0.1
regex==2023.6.3
requests==2.31.0
six==1.16.0
tiktoken==0.4.0
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2